Fixing unsigned argument type

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: pmelo-ca <pmelo-ca@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/25 14:51:05 by pmelo-ca          #+#    #+#             */
-/*   Updated: 2023/10/03 19:52:38 by pmelo-ca         ###   ########.fr       */
+/*   Updated: 2023/10/03 21:38:49 by pmelo-ca         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int	ft_putnbr_base(int nbr, char *base);
 int	ft_putchar(char c);
 int	ft_putstr(char *str);
 int	ft_putnbr(int nb);
-int	ft_put_unsigned(int nb);
+int	ft_put_unsigned(long int nb);
 int ft_put_voidpointer(char *str);
 int ft_intlen(int nb);
